pyproject.toml
[tool.poetry]
name = "signal-controller"
version = "0.1.0"
description = ""
authors = ["duyet-lfg <duyet.mai@lfglobaltech.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
